add faucet
