+++ internal/rpc/rpc.go
@@ -13,6 +13,7 @@ import (
 	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/btcsuite/btcd/rpcclient"
 	"github.com/btcsuite/btcd/wire"
+	"github.com/btcsuite/btcutil"
 )
 
 // Client is an interface that provides access to certain methods of type rpcclient.Client
@@ -20,6 +21,8 @@ type Client interface {
 	ListUnspent() ([]btcjson.ListUnspentResult, error)
 	ImportAddress(address string) error
 	SendRawTransaction(tx *wire.MsgTx, allowHighFees bool) (*chainhash.Hash, error)
+	SendToAddress(address btcutil.Address, amount btcutil.Amount) (*chainhash.Hash, error)
+	Generate(numBlocks uint32) ([]*chainhash.Hash, error)
 	// TODO: add Shutdown func
 }
